Added File Structure
Out is where compiled things go, and src is where typescript files go. TestParser is an object, and can be imported as it's done in TestTS and be used anywhere.
