+++ tsconfig.json
@@ -1,7 +1,8 @@
 {
     "compilerOptions": {
-        "target":"ES2023",
+        "target":"es6",
         "module":"CommonJS",
-        "outDir":"out",
+        "outDir":"./out",
+        "rootDir": "./src",
     }
 }
